+++ neural_speed/models/model_utils/model_types.h
@@ -302,7 +302,7 @@ struct model_context {
 
   size_t mem_per_token = 0;
 
-  int model_scratch_enlarge_scale = 1;  // model memory scratch enlarge scale
+  float model_scratch_enlarge_scale = 1.0f;  // model memory scratch enlarge scale
 
   // decode output (3-dimensional array: [batch_size] [n_tokens] [n_vocab])
   std::vector<float> logits;
@@ -423,7 +423,7 @@ struct model_context_params {
   // global generation config
   generation_config gen_conf;
   // model memory scratch enlarge scale
-  int model_scratch_enlarge_scale;
+  float model_scratch_enlarge_scale;
 
   // called with a progress value between 0 and 1, pass nullptr to disable
   model_progress_callback progress_callback;
